requirements.txt
peewee==3.16.0
psycopg2-binary==2.9.5
faust-streaming==0.10.4
gcsfs==2023.1.0
requests >= 2.28.2
fastapi >= 0.92.0
uvicorn >= 0.20.0
google-api-python-client >= 2.80.0